+++ src/OpenTelemetry.Exporter.Geneva/Metrics/GenevaMetricExporterOptions.cs
@@ -1,6 +1,8 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+#nullable enable
+
 using System.Globalization;
 using OpenTelemetry.Internal;
 
@@ -11,14 +13,14 @@ namespace OpenTelemetry.Exporter.Geneva;
 /// </summary>
 public class GenevaMetricExporterOptions
 {
-    private IReadOnlyDictionary<string, object> prepopulatedMetricDimensions;
+    private IReadOnlyDictionary<string, object>? prepopulatedMetricDimensions;
     private int metricExporterIntervalMilliseconds = 60000;
 
     /// <summary>
     /// Gets or sets the ConnectionString which contains semicolon separated list of key-value pairs.
     /// For e.g.: "Account=OTelMonitoringAccount;Namespace=OTelMetricNamespace".
     /// </summary>
-    public string ConnectionString { get; set; }
+    public string? ConnectionString { get; set; }
 
     /// <summary>
     /// Gets or sets the metric export interval in milliseconds. The default value is 60000.
@@ -41,7 +43,7 @@ public class GenevaMetricExporterOptions
     /// <summary>
     /// Gets or sets the pre-populated dimensions for all the metrics exported by the exporter.
     /// </summary>
-    public IReadOnlyDictionary<string, object> PrepopulatedMetricDimensions
+    public IReadOnlyDictionary<string, object>? PrepopulatedMetricDimensions
     {
         get
         {
@@ -67,12 +69,13 @@ public class GenevaMetricExporterOptions
                     throw new ArgumentException($"The dimension: {entry.Key} exceeds the maximum allowed limit of {GenevaMetricExporter.MaxDimensionNameSize} characters for a dimension name.");
                 }
 
-                if (entry.Value == null)
+                string? dimensionValue;
+                if (entry.Value == null
+                    || (dimensionValue = Convert.ToString(entry.Value, CultureInfo.InvariantCulture)) == null)
                 {
                     throw new ArgumentNullException($"{nameof(this.PrepopulatedMetricDimensions)}[\"{entry.Key}\"]");
                 }
 
-                var dimensionValue = Convert.ToString(entry.Value, CultureInfo.InvariantCulture);
                 if (dimensionValue.Length > GenevaMetricExporter.MaxDimensionValueSize)
                 {
                     throw new ArgumentException($"Value provided for the dimension: {entry.Key} exceeds the maximum allowed limit of {GenevaMetricExporter.MaxDimensionValueSize} characters for dimension value.");