+++ src/OpenTelemetry.SemanticConventions/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Updated to `v1.27.0` release of OpenTelemetry Semantic Conventions.
+  ([#2069](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2069))
+
+* Updated to `v1.26.0` release of OpenTelemetry Semantic Conventions.
+  ([#2040](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2040))
+
 * Updated to `1.25.0` release of OpenTelemetry Semantic Conventions.
   ([#1672](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1672))